requirements.txt
python-dotenv
mysql-connector-python
flask
flask-socketio
eventlet
geopy
requests
pyTelegramBotAPI
